chore: Remove unused function causing compilation error
The unused function 'mhi_system_resume' was causing a compilation error when compiling with Openwrt. To resolve this issue, the function has been commented out. This change follows the solution provided by Quectel in their forum.

Refactor the code to remove the commented function and update the code to enable power management during system resume.

Co-authored-by: MDSDTech